Generate audit description on render
