+++ components/audit_table/audit_row/audit_row.tsx
@@ -2,15 +2,21 @@
 // See LICENSE.txt for license information.
 
 import React from 'react';
-import {FormattedDate, FormattedTime} from 'react-intl';
+import {FormattedDate, FormattedTime, useIntl} from 'react-intl';
+
+import {useSelector} from 'react-redux';
 
 import {Audit} from '@mattermost/types/audits';
 
-import {PropsFromRedux} from '.';
+import holders from '../holders';
+import {toTitleCase} from 'utils/utils';
+import {getUser} from 'mattermost-redux/selectors/entities/users';
+import {GlobalState} from '@mattermost/types/store';
 
-interface Props extends PropsFromRedux {
+interface Props {
     audit: Audit;
-    desc: string;
+    actionURL: string;
+    desc?: string;
     showUserId: boolean;
     showIp: boolean;
     showSession: boolean;
@@ -18,12 +24,37 @@ interface Props extends PropsFromRedux {
 
 export default function AuditRow({
     audit,
-    desc,
-    userId,
+    actionURL,
+    desc: aDesc,
     showUserId,
     showIp,
     showSession,
 }: Props) {
+    const intl = useIntl();
+    let desc = aDesc;
+    if (!desc) {
+        /* Currently not called anywhere */
+        if (audit.extra_info.indexOf('revoked_all=') >= 0) {
+            desc = intl.formatMessage(holders.revokedAll);
+        } else {
+            let actionDesc = '';
+            if (actionURL && actionURL.lastIndexOf('/') !== -1) {
+                actionDesc = actionURL.substring(actionURL.lastIndexOf('/') + 1).replace('_', ' ');
+                actionDesc = toTitleCase(actionDesc);
+            }
+
+            let extraInfoDesc = '';
+            if (audit.extra_info) {
+                extraInfoDesc = audit.extra_info;
+
+                if (extraInfoDesc.indexOf('=') !== -1) {
+                    extraInfoDesc = extraInfoDesc.substring(extraInfoDesc.indexOf('=') + 1);
+                }
+            }
+            desc = actionDesc + ' ' + extraInfoDesc;
+        }
+    }
+
     const date = new Date(audit.create_at);
     const timestamp = (
         <div>
@@ -48,6 +79,8 @@ export default function AuditRow({
     const ip = audit.ip_address;
     const sessionId = audit.session_id;
 
+    const user = useSelector((state: GlobalState) => getUser(state, audit.user_id));
+    const userId = user.email ?? audit.user_id;
     let uContent;
     if (showUserId) {
         uContent = <td className='word-break--all'>{userId}</td>;
